change multiframe fusion method to log-odds
